requirements.txt
Django
pip
pytz
setuptools
sqlparse
selenium
python-dotenv
bs4
pytest-django
pytest-pythonpath
-e .
isort
flake8
requests
pytest-cov